pyproject.toml
[project]
name = "capital-bikeshare"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Jason Gardner", email = "jmgjasongardner@gmail.com"},
]
dependencies = [
    "aiobotocore~=2.13.0",
    "aiohttp~=3.9.5",
    "aioitertools~=0.11.0",
    "aiosignal~=1.3.1",
    "altair~=5.3.0",
    "argparse~=1.4.0",
    "attrs~=23.2.0",
    "beautifulsoup4~=4.12.3",
    "boto3~=1.34.106",
    "botocore~=1.34.106",
    "branca~=0.7.2",
    "certifi~=2024.2.2",
    "charset-normalizer~=3.3.2",
    "contourpy~=1.2.1",
    "cycler~=0.12.1",
    "folium~=0.16.0",
    "fonttools~=4.52.3",
    "frozenlist~=1.4.1",
    "fsspec~=2024.5.0",
    "idna~=3.7",
    "jinja2~=3.1.4",
    "jmespath~=1.0.1",
    "joblib~=1.4.2",
    "jsonschema~=4.22.0",
    "jsonschema-specifications~=2023.12.1",
    "kiwisolver~=1.4.5",
    "MarkupSafe~=2.1.5",
    "matplotlib~=3.9.0",
    "multidict~=6.0.5",
    "numpy~=1.26.4",
    "packaging~=24.0",
    "pandas~=2.2.2",
    "pillow~=10.3.0",
    "polars~=0.20.30",
    "pyarrow~=18.1.0",
    "pyparsing~=3.1.2",
    "python-dateutil~=2.9.0.post0",
    "python-dotenv~=1.0.1",
    "pytz~=2024.1",
    "referencing~=0.35.1",
    "requests~=2.32.2",
    "rpds-py~=0.18.1",
    "s3fs~=2024.5.0",
    "s3transfer~=0.10.1",
    "scikit-learn~=1.5.0",
    "scipy~=1.13.1",
    "six~=1.16.0",
    "soupsieve~=2.5",
    "threadpoolctl~=3.5.0",
    "toolz~=0.12.1",
    "tzdata~=2024.1",
    "urllib3~=2.2.1",
    "wrapt~=1.16.0",
    "xyzservices~=2024.4.0",
    "yarl~=1.9.4",
    "zipfile36~=0.1.3"
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
python = ">=3.12,<3.13"
distribution = false

[tool.pdm.build]
includes = ["src"]

[tool.pdm.dev-dependencies]
capitalbike = { path = "src", editable = true }

